WIP Big Refactor
This adds a Dreamcast struct that will hold all the components of the system and serve
as a bus (I don't emulate the different bus, tying the components this way seems alright for now).
Only the CPU needs to keep a reference to the DC for now, I think that's acceptable.
The main executable compiles, but I still have a bunch of tests to convert.