Swap register banks only when legitimate.
rb value is ignored if not in privileged mode (md == 1). I have no idea if the case rb = 1 and md = 0 actually comes
up, but I realized this could be an issue while investigating some weird behaviour in Power Stone. Turns out this was not
the issue (md is always 1 in this particular case), but this still feels right.