Order Independent Transparency. First draft.
This is really bad, but it's working. Needs a lot of code cleanups. Also we're very limited in term of memory (storage buffer are limited to around 128MB here), we should probably split up the screen when processing OIT. Blend pass should be converted to a compute shader (load and store from the same texture).