[Renderer] Monitor changes in the palette RAM to invalidate palette texture as needed.
for some reason Soulcalibur uses them before correctly setting the palette, that's pretty weird (might be a timing issue somewhere?). In any case, this fixes the character portraits in character select.
Not sure if hashing the palette RAM is the best way to do it, but it's easier than tracking the writes, and doesn't seem too costly.