[renderer] Yet another attempt at handling direct writes to the framebuffer.
This time I'm monitoring writes to VRAM and flag the framebuffer as dirty when it is written to. This works for IP.bin, but there are still false positives I do not understand, like in Speed Devils for example. It will stop rendering during loading times but keep FB_R_CTRL.enable and write something into the framebuffer area.