[renderer] Revert back to using the Holly memory directly instead of copying it to the renderer.
As expected, this was very heacy on lower end machines. This might come back to bite me if some synchronization is needed in some games, we'll see.
This also removed the list clearing from the renderer. This might actually be the cause of some issues (but I haven't verified it yet).