[renderer] Handle palettes on the GPU (#65)
Move palette texture handling to the GPU. This allows to upload only a single copy of the texture (although we're still using 4x the memory for now since it's using the same texture array as regular textures). The palette lookup is now done in a fragment/compute shader within fragment_color.
This helps a lot is some games making heavy use of palette swaps for effects.

I ran out of bits in the strip metadata, but a u16 texture index is plenty enough.

zgpu doesn't support more that 10 bindings by default. Moving to a fork for now.