[holly/renderer] Move some internal TA states from the renderer to holly.
These are colors used in intensity modes. They are now tracked in holly.zig and baked
into the VertexStrip parameters when emitted to a DisplayList.
This accomplishes two things:
 - The state is now preserved between frames, as it should.
 - Global parameters not followed by vertex data still correctly update the internal TA state.
   This case wasn't handled at all by the previous solution and the state change was dropped
   (these global parameters were never sent to the renderer).
It is also simply a more logical separation of concerns.

This fixes some graphical issues (black polygons mostly) in Power Stone and Dead or Alive 2.