[sh4] Attempt at cleaning up 64bit reads/writes.
I was investigating issues with Sega Rally 2 and started suspecting something was weird with 64bit writes. This should be closer to what the hardware does afaik, but this is actually worse in Sega Rally 2 :^) (Textures in the intro are broken in the interpreter, and behaviour differs from the JIT). Still investigating.