[renderer] Consider VO_CONTROL.pixel_double when computing global clip bounds.
Half a fix for The King of Fighters 99 Evolution. This isn't how the vertical scaling should be computed, but it should also be doubled for this game. Using VO_CONTROL.pixel_double as an heuristic for now, until I find a counter example.
Also unsure if this should affect other things than just the global clip.